quantedgeapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
bidict==0.23.1
breeze-connect==1.0.62
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
debugpy==1.8.11
decorator==5.2.1
django==4.2.20
et-xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
importlib-metadata==8.6.1
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.0.2
openpyxl==3.1.5
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==25.0
platformdirs==4.3.7
prompt-toolkit==3.0.50
psutil==5.8.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic-core==2.33.1
pydantic==2.11.2
pygments==2.19.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-engineio==4.11.2
python-socketio==5.12.1
tqdm
django-cors-headers
djangorestframework
matplotlib